Interface inicial criada

+++ app/src/main/java/com/example/app_ticket/Models/Movie.kt
@@ -5,5 +5,5 @@ data class Movie(
     val title: String,
     val poster_path: String?,
     val overview: String,
-    val vote_average: Float
+    val vote_average: Double
 )
